Move property changed reducer function to a separate file
