+++ src/store.rs
@@ -1,6 +1,7 @@
 use crate::domain::device::Device;
 use crate::domain::events::Event;
 use crate::domain::property::{BooleanProperty, NumberProperty, Property, PropertyError};
+use crate::property_changed_reducer::reduce_property_changed_event;
 use std::any::type_name;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -56,8 +57,8 @@ impl Store {
                     property_id,
                     value,
                 } => {
-                    self.reduce_property_changed_event(&device_id, &property_id, |property: &mut BooleanProperty| {
-                        return property.set_value(value).map(|_| ());
+                    reduce_property_changed_event(&mut self.devices.clone(), &device_id, &property_id, |property: &mut BooleanProperty| {
+                        property.set_value(value).map(|_| ())
                     })
                     .await;
                 }
@@ -66,55 +67,15 @@ impl Store {
                     property_id,
                     value,
                 } => {
-                    self.reduce_property_changed_event(&device_id.clone(), &property_id.clone(), move |property: &mut NumberProperty| {
-                        return property.set_value(value).map(|_| ());
-                    })
+                    reduce_property_changed_event(
+                        &mut self.devices.clone(),
+                        &device_id.clone(),
+                        &property_id.clone(),
+                        move |property: &mut NumberProperty| property.set_value(value).map(|_| ()),
+                    )
                     .await;
                 }
             }
         }
-    }
-
-    #[inline(always)]
-    async fn reduce_property_changed_event<F, T>(&mut self, device_id: &str, property_id: &str, set_value: F)
-    where
-        F: FnOnce(&mut T) -> Result<(), PropertyError>,
-        T: Property + 'static,
-    {
-        let mut write_guard = self.devices.write().await;
-
-        let Some(device) = write_guard.get_mut(device_id) else {
-            #[rustfmt::skip]
-            warn!(device_id, "⚠️ Received property changed event for unknown device '{}'", device_id);
-            return;
-        };
-
-        let Some(property) = device.properties.get_mut(property_id) else {
-            #[rustfmt::skip]
-            warn!(device_id = device.id,"⚠️ Unknown property '{}' for device '{}'", property_id, device.name);
-            return;
-        };
-
-        let previous_value = property.value_string();
-        let Some(downcast_property) = property.as_any_mut().downcast_mut::<T>() else {
-            #[rustfmt::skip]
-            warn!(device_id, "⚠️ Expected {} property for property '{}'", type_name::<T>(), &property_id);
-            return;
-        };
-
-        if let Err(err) = set_value(downcast_property) {
-            #[rustfmt::skip]
-            warn!(device_id = device.id, "⚠️ Could not set value for property '{}': {}", property_id, err);
-            return;
-        }
-
-        info!(
-            device_id = device.id,
-            "🟢 Updated device '{}', set '{}' to '{}', was '{}'",
-            device.name,
-            property.name(),
-            property.value_string(),
-            previous_value
-        );
     }
 }
